# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Forge local control API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Local-only HTTP/JSON and SSE control contract for Forge.

    This API is served over an owner-only Unix domain socket, not a TCP
    listener. The nominal server URL only supplies an HTTP authority to generic
    OpenAPI clients. M0 has no application-level authentication: the local
    process is the sole operator. Every mutation is a named command, committed
    by Core, and requires an `Idempotency-Key` header plus an expected Project
    revision. The operator is recorded in canonical Events as a `human` actor;
    M0 defines no separate local-administrator actor kind.
servers:
  - url: http://forge.local
    description: Nominal HTTP authority carried over the local Unix socket.
...
security: []
...
paths:
...
  /v1/projects/{project_id}/system-jobs:
    get:
      tags: [system-jobs]
      operationId: getProjectSystemJobStatus
      summary: Read bounded policy, jobs, onboarding receipts and aggregate attempt usage without runtime bindings.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
      responses:
        '200':
          description: Current Project SystemJob status; job list is limited to 256 records.
          content: {application/json: {schema: {$ref: '#/components/schemas/SystemJobStatusView'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/projects/{project_id}/system-jobs/{job_id}/attempts:
    get:
      tags: [system-jobs]
      operationId: listSystemJobAttempts
      summary: Read accepted-result and materialization coordinates for one Project-owned SystemJob.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: job_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: cursor, in: query, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
      responses:
        '200':
          description: Bounded attempt page; no spec, provider result body or credential binding.
          content: {application/json: {schema: {$ref: '#/components/schemas/SystemJobAttemptPage'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /v1/projects/{projectId}/employees/{employeeId}/runtime-metadata:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/EmployeeId'
    get:
      tags: [projects]
      operationId: getEmployeeRuntimeMetadata
      summary: Read safe future-Run runtime settings without prompts, secrets or host paths.
      responses:
        '200':
          description: Safe configured metadata or an explicit unavailable marker.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EmployeeRuntimeMetadata'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /v1/projects/{projectId}/tasks/{taskId}/handoffs:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/TaskId'
      - name: cursor
        in: query
        schema: {$ref: '#/components/schemas/UuidV7'}
      - name: limit
        in: query
        schema: {type: integer, minimum: 1, maximum: 20, default: 20}
    get:
      tags: [tasks]
      operationId: listTaskHandoffCoordinates
      summary: Read bounded canonical Task handoff receipts without body content.
      responses:
        '200':
          description: Task-scoped handoff coordinates and source provenance.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    maxItems: 20
                    items: {$ref: '#/components/schemas/TaskHandoffCoordinates'}
                  next_cursor: {$ref: '#/components/schemas/UuidV7'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UuidV7'
    EmployeeId:
      name: employeeId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UuidV7'
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UuidV7'
...
  responses:
    InvalidRequest:
      description: The request cannot be parsed or violates a transport contract.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: The requested resource does not exist in the Project scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unavailable:
      description: A required local Forge dependency is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    UuidV7:
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      description: A UUIDv7 identifier generated by Forge unless stated otherwise.
    Rfc3339Timestamp:
      type: string
      format: date-time
      description: UTC timestamp serialized in RFC 3339 format.
    JsonObject:
      type: object
      additionalProperties: true
      description: A JSON object interpreted only by the selected domain contract.
...
    EmployeeOnboardingStatus:
      type: object
      additionalProperties: false
      required: [employee_id, state, revision, job_id, receipt]
      properties:
        employee_id: {$ref: '#/components/schemas/UuidV7'}
        state: {type: string, enum: [pending, completed, skipped, legacy_bypass]}
        revision: {type: integer, minimum: 1, format: int64}
        job_id: {type: [string, 'null'], format: uuid}
        receipt: {type: [object, 'null'], additionalProperties: true}
    SystemJobPolicyView:
      type: object
      additionalProperties: false
      required: [enabled, max_concurrent, max_attempts_per_job, max_attempts_per_day, max_input_bytes, max_result_bytes, wall_seconds, coalesce_seconds]
      properties:
        enabled: {type: boolean}
        max_concurrent: {type: integer, minimum: 1}
        max_attempts_per_job: {type: integer, minimum: 1}
        max_attempts_per_day: {type: integer, minimum: 1}
        max_input_bytes: {type: integer, minimum: 1}
        max_result_bytes: {type: integer, minimum: 1}
        wall_seconds: {type: integer, minimum: 1}
        coalesce_seconds: {type: integer, minimum: 0}
    SystemJobView:
      type: object
      additionalProperties: false
      required: [id, project_id, kind, source_task_id, target_employee_id, generation, covered_sequence, state, reason_code]
      properties:
        id: {$ref: '#/components/schemas/UuidV7'}
        project_id: {$ref: '#/components/schemas/UuidV7'}
        kind: {type: string, enum: [onboarding, summarization]}
        source_task_id: {type: [string, 'null'], format: uuid}
        target_employee_id: {type: [string, 'null'], format: uuid}
        generation: {type: integer, minimum: 1}
        covered_sequence: {type: integer, minimum: 0}
        state: {type: string, enum: [pending, running, completed, held, cancelled]}
        reason_code: {type: [string, 'null']}
    SystemJobStatusView:
      type: object
      additionalProperties: false
      required: [policy, settings_revision, jobs, usage, onboarding]
      properties:
        policy:
          oneOf:
            - {$ref: '#/components/schemas/SystemJobPolicyView'}
            - {type: 'null'}
        settings_revision: {type: [integer, 'null'], minimum: 1}
        jobs: {type: array, maxItems: 256, items: {$ref: '#/components/schemas/SystemJobView'}}
        usage:
          type: object
          additionalProperties: false
          required: [attempts_last_24h, running, usage_status]
          properties:
            attempts_last_24h: {type: integer, minimum: 0}
            running: {type: integer, minimum: 0}
            usage_status: {type: string}
        onboarding: {type: array, items: {$ref: '#/components/schemas/EmployeeOnboardingStatus'}}
    SystemJobAttemptView:
      type: object
      additionalProperties: false
      required: [id, job_id, generation, run_id, state, result_accepted, result_message_id, created_at, completed_at, materialized_entries]
      properties:
        id: {$ref: '#/components/schemas/UuidV7'}
        job_id: {$ref: '#/components/schemas/UuidV7'}
        generation: {type: integer, minimum: 1}
        run_id: {$ref: '#/components/schemas/UuidV7'}
        state: {type: string, enum: [running, completed, stopped, failed, superseded]}
        result_accepted: {type: boolean}
        result_message_id: {type: [string, 'null'], format: uuid}
        created_at: {$ref: '#/components/schemas/Rfc3339Timestamp'}
        completed_at: {type: [string, 'null'], format: date-time}
        materialized_entries:
          type: array
          maxItems: 18
          items:
            type: object
            additionalProperties: false
            required: [id, revision]
            properties:
              id: {$ref: '#/components/schemas/UuidV7'}
              revision: {type: integer, minimum: 1}
    SystemJobAttemptPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: {type: array, maxItems: 50, items: {$ref: '#/components/schemas/SystemJobAttemptView'}}
        next_cursor: {$ref: '#/components/schemas/UuidV7'}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - validation_failed
            - not_found
            - conflict
            - stale_revision
            - idempotency_conflict
            - forbidden
            - execution_stopped
            - cursor_invalid
            - unavailable
            - internal
        message:
          type: string
        details:
          $ref: '#/components/schemas/JsonObject'
        request_id:
          $ref: '#/components/schemas/UuidV7'
...
    EmployeeRuntimeMetadata:
      oneOf:
        - type: object
          additionalProperties: false
          required: [availability, employee_id]
          properties:
            availability: {type: string, const: unavailable}
            employee_id: {$ref: '#/components/schemas/UuidV7'}
        - type: object
          additionalProperties: false
          required: [availability, employee_id, binding_revision, updated_at, profile_id, profile_revision, adapter_id, adapter_version, provider_id, model, credential_binding_id, credential_delivery, image_digest, surface_kind, access, limits, budget, prompts_available]
          properties:
            availability: {type: string, const: available}
            employee_id: {$ref: '#/components/schemas/UuidV7'}
            binding_revision: {type: integer, minimum: 1}
            updated_at: {$ref: '#/components/schemas/Rfc3339Timestamp'}
            profile_id: {$ref: '#/components/schemas/UuidV7'}
            profile_revision: {type: integer, minimum: 1}
            adapter_id: {type: string}
            adapter_version: {type: string}
            provider_id: {type: string}
            model: {type: string}
            credential_binding_id: {$ref: '#/components/schemas/UuidV7'}
            credential_delivery: {type: string, enum: [proxy_only, isolated_runtime_secret, trusted_host]}
            image_digest: {type: string, pattern: '^[0-9a-f]{64}$'}
            surface_kind: {type: string, enum: [none, filesystem_sandbox, git_worktree, git_unborn, git_unborn_candidate_snapshot, git_candidate_snapshot]}
            access: {type: string, enum: [read_write, read_only]}
            limits:
              type: object
              additionalProperties: false
              required: [cpu_millis, memory_bytes, pids, wall_seconds, stop_grace_seconds]
              properties:
                cpu_millis: {type: integer, minimum: 1}
                memory_bytes: {type: integer, minimum: 1}
                pids: {type: integer, minimum: 1}
                wall_seconds: {type: integer, minimum: 1}
                stop_grace_seconds: {type: integer, minimum: 1}
            budget:
              type: object
              additionalProperties: false
              required: [max_output_bytes, requests_per_minute, tokens_per_minute, max_spend_microusd]
              properties:
                max_output_bytes: {type: integer, minimum: 1}
                requests_per_minute: {type: integer, minimum: 1}
                tokens_per_minute: {type: integer, minimum: 1}
                max_spend_microusd: {type: [integer, 'null'], minimum: 1}
            prompts_available: {type: boolean, const: true}
...
    TaskHandoffCoordinates:
      type: object
      additionalProperties: false
      required: [id, project_id, task_id, kind, source_kind, source_id, created_at, content_availability]
      properties:
        id: {$ref: '#/components/schemas/UuidV7'}
        project_id: {$ref: '#/components/schemas/UuidV7'}
        task_id: {$ref: '#/components/schemas/UuidV7'}
        kind: {type: string, enum: [accepted, interrupted]}
        source_kind: {type: string, enum: [run, hook, git_integration, command]}
        source_id: {$ref: '#/components/schemas/UuidV7'}
        created_at: {$ref: '#/components/schemas/Rfc3339Timestamp'}
        content_availability: {type: string, const: unavailable}
...